Refactor: Improve code structure and logging
This commit refactors the netcode modules for better organization and logging. It also updates ICE servers and improves desync detection logic.
