+++ src/netcode/rollback-netcode.js
@@ -8,7 +8,7 @@
  * - Adaptive performance tuning based on network conditions
  */
 
-import { createLogger } from '../logger.js'
+import { createLogger } from '../utils/logger.js'
 
 // Constants for rollback system
 const MAX_ROLLBACK_FRAMES = 12 // Increased for better recovery
@@ -169,7 +169,9 @@ class RollbackNetcode {
    * Add a remote player to the session with enhanced tracking
    */
   addPlayer(playerId, inputDelay = INPUT_DELAY_FRAMES, connectionInfo = {}) {
-    if (this.players.has(playerId)) {return}
+    if (this.players.has(playerId)) {
+      return
+    }
     
     this.players.set(playerId, {
       id: playerId,
@@ -217,7 +219,9 @@ class RollbackNetcode {
    * Start the game loop
    */
   start() {
-    if (this.running) {return}
+    if (this.running) {
+      return
+    }
     
     this.running = true
     this.lastFrameTime = performance.now()
@@ -236,7 +240,9 @@ class RollbackNetcode {
    * Main game loop with fixed timestep
    */
   gameLoop() {
-    if (!this.running) {return}
+    if (!this.running) {
+      return
+    }
     
     const now = performance.now()
     const deltaTime = now - this.lastFrameTime
@@ -893,9 +899,9 @@ class RollbackNetcode {
       return 'fair'
     } else if (avgLatency > 50) {
       return 'good'
-    } 
+    } else {
       return 'excellent'
-    
+    }
   }
   
   /**
